Build LLVM properly
Build llvm with libc++ as the standard. In order to do this, one needs
to build the LLVM suite in the following order:

1. LLVM/
        clang

2. LLVM/
        libcxx
        libcxxabi

Using the installed compiler above as a bootstrap compiler, it is then
possible to build the entire LLVM suite all at once:

3. LLVM/
        clang
        libcxx
        libcxxabi
        compiler-rt
        openmp
        lldb

VERY IMPORTANT NOTE: It was found necessary to build each above step, using
a new cloned project directory (or rather, when moving on to step 2, you must
delete the clang project).

Note: The influential cmake arguments that makes the stdlib=libc++ used by
default was:

  -DCLANG_DEFAULT_CXX_STDLIB="libc++"
  -DLLVM_ENABLE_LIBCXX="ON"
  -DLLVM_ENABLE_LIBCXXABI="ON"

...or so I believe. Building LLVM still feels like voodoo.